rename "api-client.ts" to "zebra-client.ts"
